+++ platform/osx/test/MGLOfflineStorageTests.m
@@ -30,7 +30,7 @@
 - (void)testAddPack {
     NSUInteger countOfPacks = [MGLOfflineStorage sharedOfflineStorage].packs.count;
     
-    NSURL *styleURL = [MGLStyle lightStyleURL];
+    NSURL *styleURL = [MGLStyle lightStyleURLWithVersion:8];
     /// Somewhere near Grape Grove, Ohio, United States.
     MGLCoordinateBounds bounds = {
         { 39.70358155855172, -83.69506472545841 },